// src/app/types/Restypes.tsx
export interface SubscriptionDetails {
  user_id: number;
  plan_name: string;
  price: string;
  expire_date: Date;
  plan_image: string;
}

//  Profile Update

export interface ProfileUpdate {
  status: boolean;
  message: string;
  userdetails: Userdetails;
  subscriptionDetails: SubscriptionDetails;
}

export interface SubscriptionDetails {
  user_id: number;
  plan_name: string;
  price: string;
  expire_date: Date;
  plan_image: string;
}

export interface Userdetails {
  id: number;
  first_name: string;
  last_name: string;
  mobile: string;
  email: string;
  password: string;
  username: string;
  country_code: string;
  role: string;
  image: string;
  subscribed_user: number;
}